putting common code in base class
